Deduplicate transaction modal css builder
